requirements.txt
asgiref==3.2.7
astral==2.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
Brotli==1.0.9
cffi==1.14.3
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.4.0
django-multiselectfield==0.1.12
django-mysql==3.9.0
djangorestframework==3.12.1
dpd-components==0.1.0
flake8==3.8.4
Flask==1.1.2
Flask-Compress==1.6.0
future==0.18.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mysql-client==0.0.1
mysqlclient==2.0.1
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.3.1
txaio==20.4.1
Werkzeug==1.0.1
